Use Minitest module instead of MiniTest compatibility layer
The Minitest package has long been capitalized like that. The variant
MiniTest is retained only as a compatibility layer. Switch to the
current module name.

This change is motivated by Fedora, which already uses minitest 5.19.
That version requires setting the env variable MT_COMPAT to import the
compatibility module, and thus revealed the problem in Fedora build
system.

Signed-off-by: Otto Liljalaakso <otto.liljalaakso@iki.fi>